requirements.txt
albumentations==1.2.1
backcall==0.2.0
beautifulsoup4==4.9.3
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
conda==4.10.1
conda-build==3.21.4
conda-package-handling==1.7.3
cryptography==3.4.7
cycler==0.11.0
decorator==5.0.9
dnspython==2.1.0
efficientnet-pytorch==0.7.1
filelock==3.0.12
fonttools==4.37.1
glob2==0.7
idna==2.10
imageio==2.21.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==3.0.0
joblib==1.1.0
kiwisolver==1.4.4
libarchive-c==2.9
MarkupSafe==2.0.1
matplotlib==3.5.3
mkl-fft==1.3.0
mkl-random==1.2.1
mkl-service==2.3.0
munch==2.5.0
networkx==2.6.3
numpy==1.20.2
olefile==0.46
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.3.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==21.1.1
pkginfo==1.7.0
pretrainedmodels==0.7.4
prompt-toolkit==3.0.17
protobuf==3.20.1
psutil==5.8.0
ptyprocess==0.7.0
pycosat==0.6.3
pycparser==2.20
Pygments==2.9.0
pyOpenSSL==19.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-etcd==0.4.5
pytz==2021.1
PyWavelets==1.3.0
PyYAML==5.4.1
qudida==0.0.4
requests==2.24.0
ruamel-yaml-conda==0.15.100
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
segmentation-models-pytorch==0.3.0
setuptools==52.0.0.post20210125
six==1.15.0
soupsieve==2.2.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.12
torch==1.8.1
torchelastic==0.2.0
torchtext==0.9.1
torchvision==0.9.1
tqdm==4.51.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.2.5
wheel==0.35.1
